Update pricing parser and API to handle new data sources
The old pricing-on-demand-instances.json has been replaced by a file per OS. To support this the whole handling of downloading and merging the pricing data had to change, but at the same time support for RHEL and SLES is added.

Now an instance type will be included if it has a price, not only if it has data.