Update focus when changing element contents
When ngRepeat gets its items in a different order, it recreates all the
underlying elements. Given that there are not that many instance types,
this in itself isn't so bad. However, focus is lost from any focused
element, which is a problem for the calculator input fields.

This introduces a solution from http://stackoverflow.com/a/18295416 to
re-focus an element that was reordered due to a content change. The
element is still dropped and a new one created, but for most uses it
appears as if the element was just moved.