Load reserved rates & costs at the same time as the on demand rate
You want to see reserved pricing along with on demand pricing to compare them. This makes a self join of the cost table to get the prices side by side. This requires knowing the purchase option, lease contract length and offering class of on demand instances, but they should always be empty strings.

The UI no longer needs to provide the option to select the on demand combination of purchase option, lease contract length and offering class, and I'm also removing some of the other filter options that are not relevant (maybe these should be filtered out when importing).

The query seems to take significantly longer to run, and I'm getting some memory errors from SQL.js intermittently, so maybe it would be a better idea to do two queries and join the results in the application instead. I assume the problem has to do with memory pressure when doing the large join, or something like that.