Refactor the filtering in App
Change from having constants with different configs and multiple reactive properties with different parts to having a more complex model in a single reactive property. Each filter is now an object that contains the config, default option, selected option and all options. The options have an `id` and `value` property instead of the value properties all being called different thing.

This makes it so much easier to work with the filters.

The model of each filter selector is the `id` property of the filter's `selected` property, which means that it gets overwritten when the selection changes, but the `value` property does not. Therefore a watch is needed to supply the value for `value`.

The watch for the operating system and the special handling of Windows needed a tweak to since the computed property seems to be recomputed after the watch triggers now.