Make the rows in the instance types table selectable
This required moving the instance types to the data section of the App component, because computed things are not reactive, it seems.

The instance types table does some funky stuff to make the selection expand out on the sides of the table because I want some padding to the left and right, but still have the table align with the filters.