Add a feature that shows rates by day, week, etc. in addition to hour
This adds a filter selector whose options are not loaded form the database, but instead has hard coded options.

It also adds a prop to the instances table that is used to multiply the rates. I added some missing tests around n/a as well since that broke when I started multiplying the numbers.