Replace the selects with a component that allows groups of selections
Selector shows a drop down menu with multiple groups of options, so all the reservation options can be in one drop down, and all the software options in another. This makes the UI a bit cleaner. Most of the time you won't care about the different reservation options, for example, and you'll almost never care about the different options that are only available when you choose Windows.

The component is pretty ugly as of now, unfortunately.