Work around an error in the 2017-04-21 EC2 prices
There is an inconsistent row for f1.2xlarge that has TermType "OnDemand" but at the same time a PriceDescription that says "ReservedHostUsage:f1". There is an otherwise identical row that says "BoxUsage:f1.2xlarge" in the PriceDescription. They both have the same LeaseContractLength, PurchaseOption and OfferingClass and there is not upfront cost row. I'm pretty sure this is an error.